+++ test/Test/Fixtures/progress/macro-rescan-repeated-name.hs
@@ -1,0 +1,3 @@
+#define F(x) (x)
+#define G(x) F(x) + F(F(x))
+h = G(3)